Double went roundable I need not roundable number.

+++ src/main/java/uz/backall/profit/ProfitController.java
@@ -74,7 +74,7 @@ public class ProfitController {
   }
 
   @GetMapping("/month/amount")
-  public ResponseEntity<Double> getThisMonthAmountInfo(
+  public ResponseEntity<Long> getThisMonthAmountInfo(
     @RequestParam(value = "storeId") Long storeId
   ) {
     return ResponseEntity.ok(
