+++ src/main/java/uz/backall/profit/profitAmountDate/ProfitAmountDateService.java
@@ -137,7 +137,7 @@ public class ProfitAmountDateService {
     );
   }
 
-  public Double getMonthAmount(Long storeId) {
+  public Long getMonthAmount(Long storeId) {
     Calendar calendar = Calendar.getInstance();
     int monthNumber = calendar.get(Calendar.MONTH) + 1; // Months are 0-based in Calendar
     int year = calendar.get(Calendar.YEAR);
@@ -145,7 +145,7 @@ public class ProfitAmountDateService {
     // Construct the date pattern for the given month and year (e.g., "2024-08-%")
     String datePattern = year + "-" + (monthNumber < 10 ? "0" : "") + monthNumber + "-%";
 
-    Double totalAmount = profitAmountDateRepository.findTotalAmountByStoreIdAndDatePattern(storeId, datePattern);
-    return totalAmount != null ? totalAmount : 0.0;
+    Long totalAmount = profitAmountDateRepository.findTotalAmountByStoreIdAndDatePattern(storeId, datePattern);
+    return totalAmount != null ? totalAmount : 0;
   }
 }
