+++ src/main/java/uz/backall/sell/sellAmountDate/SellAmountDateRepository.java
@@ -22,5 +22,5 @@ public interface SellAmountDateRepository extends JpaRepository<SellAmountDateEn
   Optional<SellAmountDateEntity> findByStoreIdAndDate(Long storeId, String date);
 
   @Query("SELECT SUM(s.amount) FROM SellAmountDateEntity s WHERE s.storeId = :storeId AND s.date LIKE :datePattern")
-  Double findTotalAmountByStoreIdAndDatePattern(@Param("storeId") Long storeId, @Param("datePattern") String datePattern);
+  Long findTotalAmountByStoreIdAndDatePattern(@Param("storeId") Long storeId, @Param("datePattern") String datePattern);
 }
